firts work version

+++ Spv.WinForms/Models/Dict/H264.cs
@@ -4,9 +4,9 @@ public class H264
 {
     public string ValveType { get; set; } = string.Empty;
     public string Pressure { get; set; } = string.Empty;
-    public int Vessel { get; set; }
-    public float Time { get; set; }
-    public float Rate { get; set; }
+    public double Vessel { get; set; }
+    public double Time { get; set; }
+    public double Rate { get; set; }
     public string Bemerkungen { get; set; } = string.Empty;
     public string ModifiedBy { get; set; } = string.Empty;
 }
